requirements.txt
mysql-connector-python==8.0.33
tkcalendar==1.6.1
